Comment out Lifetime property and related code
The Lifetime property and its XML documentation have been commented out in both EvDbStreamFactoryAttribute<TEventType, TTopicProducer> and EvDbStreamFactoryAttribute<TEventType> classes. In the EvDbGenerator class, the code that was retrieving the Lifetime property has been commented out and replaced with a constant value "Singleton". Additionally, the Lifetime parameter in the EvDbStreamFactory attribute in the SchoolStreamFactory class has been commented out.