Fix/ci 2025 01 16 (#183)
* Add .NET setup step to workflow

Modified `tagging_message` line to remove trailing space and re-added it without changes. Added a new step "Setup .NET" using `actions/setup-dotnet@v4` to configure .NET version `8.0.x` in the workflow.

* Update package version to 1.1.149

Updated the version number in Directory.Build.props from 1.1.146 to 1.1.149. This change is reflected in both the <Version> and <PackageVersion> elements, ensuring the new version is correctly identified and used in the build process.