Add guistack service and enhance health checks
This commit introduces a new `guistack` service in the `docker-compose.yml` file, which depends on `localstack`. Health checks have been added for `zookeeper`, `kafka`, `schema-registry`, `kafdrop`, and `kafka-ui` to ensure proper service readiness. The `kafka-topics-generator` now waits for Kafka to be ready before creating topics. The `prometheus` service is updated to a specific version with health checks, and the `grafana` service includes a health check and an admin password. Additionally, the `prometheus.yml` configuration has been improved with new scrape configurations for `kafka`, `kafka-connect`, and `schema-registry`, along with global scrape settings.